requirements.txt
anthropic
streamlit
python-dotenv
requests
python-docx
pytest
